Reorganised lib and content
