+++ guide/content/radios.slim
@@ -5,17 +5,17 @@ h2.govuk-heading-l From a collection
 == render('/partials/example-fig.*',
   caption: "Radio buttons collection with no additional params",
   code: radio_field,
-  sample_data: departments_data)
+  sample_data: departments_data_raw)
 
 == render('/partials/example-fig.*',
   caption: "Radio buttons collection with legend and hint text",
   code: radio_field_with_legend_and_hint,
-  sample_data: departments_data)
+  sample_data: departments_data_raw)
 
 == render('/partials/example-fig.*',
   caption: "Radio buttons collection with descriptions",
   code: radio_field_with_label_and_descriptions,
-  sample_data: lunch_options)
+  sample_data: lunch_options_raw)
 
 h2.govuk-heading-l Supplying a custom list of options
