+++ guide/content/select.slim
@@ -2,17 +2,17 @@ h1.govuk-heading-xl Select field
 
 p.govuk-body-l
 
-== render_figure('/partials/example-fig.*',
+== render('/partials/example-fig.*',
   caption: "Select field with no additional params",
   code: select_field,
-  sample_data: departments_data)
+  sample_data: departments_data_raw)
 
-== render_figure('/partials/example-fig.*',
+== render('/partials/example-fig.*',
   caption: "Select field with a label",
   code: select_field_with_label,
-  sample_data: departments_data)
+  sample_data: departments_data_raw)
 
-== render_figure('/partials/example-fig.*',
+== render('/partials/example-fig.*',
   caption: "Select field with hint text",
   code: select_field_with_hint_text,
-  sample_data: departments_data)
+  sample_data: departments_data_raw)
