+++ guide/content/text-field.slim
@@ -37,22 +37,22 @@ ul.govuk-list
       | https://html.spec.whatwg.org/multipage/input.html#states-of-the-type-attribute
 
 
-== render_figure('/partials/example-fig.*',
+== render('/partials/example-fig.*',
   caption: "Generating a text input with no additional params",
   code: text_field)
 
-== render_figure('/partials/example-fig.*',
+== render('/partials/example-fig.*',
   caption: "Generating a text input with a label",
   code: text_field_with_label)
 
-== render_figure('/partials/example-fig.*',
+== render('/partials/example-fig.*',
   caption: "Generating a text input with some hint text",
   code: text_field_with_hint)
 
-== render_figure('/partials/example-fig.*',
+== render('/partials/example-fig.*',
   caption: "Generating a text input with some extra attributes",
   code: text_field_with_attributes)
 
-== render_figure('/partials/example-fig.*',
+== render('/partials/example-fig.*',
   caption: "Generating a text input with a custom width",
   code: text_field_with_custom_width)
