+++ guide/lib/examples/select.rb
@@ -1,15 +1,5 @@
 module Examples
   module Select
-    def departments_data
-      <<~DATA
-        @departments = [
-          OpenStruct.new(id: 1, name: 'Sales'),
-          OpenStruct.new(id: 2, name: 'Marketing'),
-          OpenStruct.new(id: 3, name: 'Finance')
-        ]
-      DATA
-    end
-
     def select_field
       <<~SNIPPET
         = f.govuk_collection_select :department_id,
