+++ guide/lib/helpers.rb
@@ -9,11 +9,12 @@ require 'govuk_design_system_formbuilder'
 
 use_helper Nanoc::Helpers::Rendering
 use_helper Nanoc::Helpers::LinkTo
-use_helper BuilderExamples
-use_helper Formatters
-use_helper TextHelpers
+use_helper Helpers::Formatters
+use_helper Helpers::TextHelpers
 use_helper Examples::TextInput
 use_helper Examples::Select
 use_helper Examples::Radios
+use_helper Setup::FormBuilderObjects
+use_helper Setup::ExampleData
 
 class TestHelper < ActionView::Base; end
