+++ guide/lib/helpers/person.rb
