+++ guide/lib/setup/form_builder_objects.rb
@@ -1,0 +1,16 @@
+module Setup
+  module FormBuilderObjects
+    def builder
+      GOVUKDesignSystemFormBuilder::FormBuilder.new(:person, object, helper, {})
+    end
+    alias_method :f, :builder
+
+    def object
+      Person.new
+    end
+
+    def helper
+      TestHelper.new
+    end
+  end
+end
