--- guide/lib/text_helpers.rb
@@ -1,13 +1,0 @@
-module TextHelpers
-  def documentation_link
-    'https://www.rubydoc.info/gems/govuk_design_system_formbuilder/GOVUKDesignSystemFormBuilder/Builder'
-  end
-
-  def github_link
-    'https://github.com/DFE-Digital/govuk_design_system_formbuilder'
-  end
-
-  def design_system_link
-    'https://design-system.service.gov.uk'
-  end
-end
